python-service/google_docs_api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
pydantic==2.5.3
